requirements.txt
pandas
openpyxl
pyyaml
python-dotenv
google-ads
facebook-business
